requirements.txt
feedparser>=6.0.0
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
httpx>=0.25.0
